+++ plugins/source_validate_test.go
@@ -1,10 +1,10 @@
 package plugins
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/google/go-cmp/cmp"
 )
 
 func TestSourcePlugin_listAndValidateAllResources(t *testing.T) {
@@ -18,88 +18,193 @@ func TestSourcePlugin_listAndValidateAllResources(t *testing.T) {
 	}{
 		{
 			name:                "should return all tables when '*' is provided",
-			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}, {Name: "table 3"}}},
+			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}, {Name: "table3"}}},
 			configurationTables: []string{"*"},
-			want:                []string{"table 1", "table 2", "table 3"},
+			want:                []string{"table1", "table2", "table3"},
 			wantErr:             false,
 		},
 		{
 			name:                    "should return all tables when '*' is provided, excluding skipped tables",
-			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}, {Name: "table 3"}}},
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}, {Name: "table3"}}},
 			configurationTables:     []string{"*"},
-			configurationSkipTables: []string{"table 1", "table 3"},
-			want:                    []string{"table 2"},
+			configurationSkipTables: []string{"table1", "table3"},
+			want:                    []string{"table2"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should return error when invalid table is skipped",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}, {Name: "table3"}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{"table4"},
+			wantErr:                 true,
+		},
+		{
+			name: "should return the parent and all its descendents",
+			plugin: SourcePlugin{tables: []*schema.Table{
+				{Name: "other_table", Relations: []*schema.Table{
+					{Name: "other_sub_table", Parent: &schema.Table{Name: "main_table"}}},
+				},
+				{Name: "main_table", Relations: []*schema.Table{
+					{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}, Relations: []*schema.Table{
+						{Name: "sub_sub_table", Parent: &schema.Table{Name: "sub_table"}},
+					}}}},
+			},
+			},
+			configurationTables:     []string{"main_table"},
+			configurationSkipTables: []string{},
+			want:                    []string{"main_table", "sub_table", "sub_sub_table"},
+			wantErr:                 false,
+		},
+		{
+			name: "should return the parent and all its descendents, but skip children and their descendents if they are skipped",
+			plugin: SourcePlugin{tables: []*schema.Table{{Name: "main_table", Relations: []*schema.Table{
+				{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}, Relations: []*schema.Table{
+					{Name: "sub_sub_table", Parent: &schema.Table{Name: "sub_table"}},
+				}}}},
+			},
+			},
+			configurationTables:     []string{"main_table"},
+			configurationSkipTables: []string{"sub_table"},
+			want:                    []string{"main_table"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should return only parent table if child table is skipped",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "main_table", Relations: []*schema.Table{{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}}}}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{"sub_table"},
+			want:                    []string{"main_table"},
 			wantErr:                 false,
 		},
 		{
 			name:                "should return specific tables when they are provided",
-			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}, {Name: "table 3"}}},
-			configurationTables: []string{"table 1"},
-			want:                []string{"table 1"},
+			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}, {Name: "table3"}}},
+			configurationTables: []string{"table1"},
+			want:                []string{"table1"},
 			wantErr:             false,
 		},
 		{
-			name:                "should return error when '*' is provided with other tables",
-			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}, {Name: "table 3"}}},
-			configurationTables: []string{"table 1", "*"},
-			wantErr:             true,
-		},
-		{
 			name:    "should return an error when nil is provided",
-			plugin:  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}}},
+			plugin:  SourcePlugin{tables: []*schema.Table{{Name: "table1"}}},
 			wantErr: true,
 		},
 		{
 			name:                    "should return an error if glob-matching is attempted in tables",
-			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}}},
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
 			configurationTables:     []string{"table*"},
 			configurationSkipTables: []string{""},
 			wantErr:                 true,
 		},
 		{
-			name:                    "should return an error if glob-matching is attempted in skipped tables",
-			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}}},
-			configurationTables:     []string{"table 1"},
-			configurationSkipTables: []string{"table *"},
-			wantErr:                 true,
+			name:                    "should return tables matching glob pattern",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"table*"},
+			configurationSkipTables: []string{"table2"},
+			want:                    []string{"table1"},
+			wantErr:                 false,
 		},
 		{
-			name:                    "should return an error when included table is skipped",
-			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}}},
-			configurationTables:     []string{"table 2", "table 1"},
-			configurationSkipTables: []string{"table 1"},
-			wantErr:                 true,
+			name:                    "should not return an error when included table is skipped",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"table2", "table1"},
+			configurationSkipTables: []string{"table1"},
+			want:                    []string{"table2"},
+			wantErr:                 false,
 		},
 		{
 			name:                    "should return an error if table is unmatched",
-			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}}},
-			configurationTables:     []string{"table 2"},
-			configurationSkipTables: []string{"table 1"},
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}}},
+			configurationTables:     []string{"table2"},
+			configurationSkipTables: []string{"table1"},
 			wantErr:                 true,
 		},
 		{
-			name:                "should return an error if child table is specified",
+			name:                "should return an error if child table is specified without its ancestors",
 			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "main_table", Relations: []*schema.Table{{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}}}}}},
 			configurationTables: []string{"sub_table"},
 			wantErr:             true,
 		},
 		{
-			name:                "should return an error if both child and parent tables are specified",
+			name:                "should return both tables if child and parent tables are specified",
 			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "main_table", Relations: []*schema.Table{{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}}}}}},
 			configurationTables: []string{"main_table", "sub_table"},
-			wantErr:             true,
+			want:                []string{"main_table", "sub_table"},
+			wantErr:             false,
+		},
+		{
+			name:                    "should return table only once, even if it is matched by multiple rules",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"*", "table2", "table1", "table*"},
+			configurationSkipTables: []string{"table2"},
+			want:                    []string{"table1"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should match prefix globs",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"*2"},
+			configurationSkipTables: []string{},
+			want:                    []string{"table2"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should match suffix globs",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"table*"},
+			configurationSkipTables: []string{},
+			want:                    []string{"table1", "table2"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should match middle globs",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"t*b*1"},
+			configurationSkipTables: []string{},
+			want:                    []string{"table1"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should skip globs",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{"t*1"},
+			want:                    []string{"table2"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should skip multiple globs",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table1"}, {Name: "table2"}, {Name: "table3"}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{"t*1", "t*2"},
+			want:                    []string{"table3"},
+			wantErr:                 false,
+		},
+		{
+			name:                    "should glob match against child tables",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "main_table", Relations: []*schema.Table{{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}}}}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{},
+			want:                    []string{"main_table", "sub_table"},
+			wantErr:                 false,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := tt.plugin.listAndValidateTables(tt.configurationTables, tt.configurationSkipTables)
+			gotTables, err := tt.plugin.listAndValidateTables(tt.configurationTables, tt.configurationSkipTables)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("SourcePlugin.listAndValidateTables() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("SourcePlugin.listAndValidateTables() = %v, want %v", got, tt.want)
+			if tt.wantErr {
+				return
+			}
+			gotNames := make([]string, len(gotTables))
+			for i := range gotTables {
+				gotNames[i] = gotTables[i].Name
+			}
+			if diff := cmp.Diff(gotNames, tt.want); diff != "" {
+				t.Errorf("SourcePlugin.listAndValidateTables() diff (+got, -want): %v", diff)
 			}
 		})
 	}